Cleanup deprecated APIs and warnings
